requirements.txt
python-dotenv==0.20.0
redis==4.3.4